+++ app/models/openstax/accounts/group_nesting.rb
@@ -38,13 +38,15 @@ module OpenStax::Accounts
     end
 
     def create_openstax_accounts_group_nesting
-      return false unless requestor
+      return false unless requestor         # Fail if requestor not set
+      return if requestor.access_token.nil? # Ignore call to Accounts if requestor is temp
 
       OpenStax::Accounts::Api.create_group_nesting(requestor, self)
     end
 
     def destroy_openstax_accounts_group_nesting
-      return false unless requestor
+      return false unless requestor         # Fail if requestor not set
+      return if requestor.access_token.nil? # Ignore call to Accounts if requestor is temp
 
       OpenStax::Accounts::Api.destroy_group_nesting(requestor, self)
     end
